backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pandas==2.0.2
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6